+++ packages/plantuml-renderer/src/index.js
@@ -1,3 +1,7 @@
 export { PlantUmlRendererError } from './errors.js';
 export { createPlantUmlRenderer } from './renderer.js';
 export { plantUmlRendererLimits } from './limits.js';
+export {
+  parsePlantUmlStandardReport,
+  sanitizePlantUmlDiagnostics,
+} from './standard-report.js';